Removed unused graphics code & moved the Mesh to the lsystem folder. Removed old shader class because we moved to PyOpenGL's convenience methods.
